Animations and dialogue trial
MariamAnimations contain the player itself. So if you've made code changes you need to put them here. I dont know how to extract just the animations. But this package contains the gameobject and in that mariam gameobject contains all her animations.
LavosBig and Small are called Beast2 and Beast respectively. They contain the animations made for the battle.
Dialouge manager goes under canvas. The code is compatible with what seer has that has his/her dialogue. When player is next to seer and presses space, the dialogue will show up.